# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic E-Commerce & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get my orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"

  # ── Admin ────────────────────────────────────────────────────────
  /admin/orders:
    get:
      operationId: getAdminOrders
      tags: [admin]
      summary: Get all orders (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/analytics:
    get:
      operationId: getAnalytics
      tags: [admin]
      summary: Get analytics (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
  /admin/low-stock:
    get:
      operationId: getLowStock
      tags: [admin]
      summary: Get low-stock products (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: threshold
          required: false
          schema:
            type: integer
            default: 5
          description: Stock level below which a product is considered low-stock
      responses:
        "200":
          description: Low-stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ── ERP: Employees ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        price:
          type: string
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        categoryId:
          type: integer
          nullable: true
        rating:
          type: string
        reviewCount:
          type: integer
        createdAt:
          type: string
      required: [id, nameAr, nameEn, price, stock, rating, reviewCount]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        totalAmount:
          type: string
        discountAmount:
          type: string
        couponCode:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, customerPhone, customerAddress, status, totalAmount]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  quantity:
                    type: integer
                  unitPrice:
                    type: string
                  product:
                    type: object
                    properties:
                      id:
                        type: integer
                      nameAr:
                        type: string
                      nameEn:
                        type: string
                      imageUrl:
                        type: string
                        nullable: true
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        couponCode:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required: [productId, quantity]
      required: [customerName, customerPhone, customerAddress]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
      required: [status]
    Analytics:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        pendingOrders:
          type: integer
        dailySales:
          type: array
          items:
            type: object
        topProducts:
          type: array
          items:
            type: object
        lowStock:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [totalOrders, totalRevenue, netProfit, pendingOrders]

    # ERP
...